requirements.txt
Flask
Pillow
tensorflow
transformers
huggingface_hub
numpy
